requirements.txt
aiofiles==0.6.0
aniso8601==8.1.1
attrs==20.3.0
click==7.1.2
fastapi==0.63.0
graphene==3.0b7
graphene-pydantic==0.2.0
graphql-core==3.1.3
graphql-relay==3.1.0
h11==0.12.0
iniconfig==1.1.1
Jinja2==2.11.3
MarkupSafe==1.1.1
neo4j-driver==4.1.1
neobolt==1.7.17
neomodel==4.0.2
packaging==20.9
pluggy==0.13.1
promise==2.3
py==1.10.0
pydantic==1.8.1
pyparsing==2.4.7
pytest==6.2.2
python-dotenv==0.17.1
pytz==2021.1
Rx==1.6.1
Shapely==1.7.1
six==1.15.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
uvicorn==0.13.4
